Fix #257: specify bridge to avoid losing traffic from the outside world
Previously, the bridge generated by flannel would be cni0 if launched under the
coreos stage1, and kvm-cni0 if launched under the kvm stage1. Since both had
the same subnet bound, this would cause a conflict where traffic could only be
delivered to one of the two.

This change specifies that the bridge should always be homeworld-cni0, so that
the two stage1s will agree on the CNI bridge to use.