Fix #120: remove resource limits on the sidecar container
Due to a low CPU limit, the sidecar container would consistently fail to start
up before at least a few minutes had passed, at which point the entire pod
would fail and get restarted due to failing liveness tests.

This is not the correct long-term solution, since there should be limits, but
it solves the immediate problem, and we'll need to generate/regenerate limits
later anyway.