glass: add new build tool
Our existing build system is complicated and very repetitive. Using
it requires specialized knowledge, and it stands in the way of new
contributors to the project getting started.

This new build tool can be used to build any of the existing
packages, once they're migrated to it.

This system takes yaml-described build processes and uses them to
perform builds. These scripts use a specialized language that more
directly expresses the kinds of builds that we need to do than the
existing system of bash scripts and debian package description files.

Although the yaml descriptions are not as concise as standard bash
scripts, they are more expressive for our uses, and should be fairly
self-describing when read, unlike our existing system that requires
understanding how multiple different scripts fit together.

This new build tool also incorporates the ability to upload new
versions of the binaries to a Google Cloud Storage bucket, where they
(in theory) can be used to launch a Hyades cluster. This is intended
to replace the existing web.mit.edu upload pattern, and act as a more
long-term solution. Other providers, such as Amazon S3 and Hyades
itself, are planned for future support.

This build system also unifies the build processes for containers,
build helpers, and debian packages, meaning that they will be able to
be merged into a unified component build tree in the future.

There is still considerable room for improvement in this system, as a
number of key functions are still only expressible through bash
scripts.

In a perfect world, I would include a more detailed explanation here
of exactly why everything about glass works the way it does, but it's
not worth my time to attempt to reconstruct each interstitial version
of the glass system.