chroot: pre-populate /fstemp/ directory
Glass attempts to auto-create the /fstemp/ folder, in which a build
can be launched if it requires more temporary working space than is
normally available on a tmpfs.

However, glass often may not have sufficient access to create
/fstemp/ when running as a normal user within the chroot. This change
autocreates the folder, so that glass will be able to use it.