ci: fix upstream verification on circleci
Circle CI uses Ubuntu 14.04, which is missing the latest version of
tar, which is needed for the --sort=name option needed for
reproducible tarballs. By running verification in the chroot, the
correct tar version is known to be available.

The upstream verification is now also reordered after the chroot init
in order to take advantage of the chroot, and we kill gpg-agent after
completing the verification process so that the 14.04 chroot variant
can safely unmount the homeworld bindmount.