debian: fix fakechroot-generated symlinks
Implement an extra build step for debian containers that fixes
fakechroot's malformed host-side symlinks.

Under fakechroot, a symlink to an absolute path within a chroot will
be generated as an absolute path in the host filesystem. For example,
a link like this:

    /lib64/ld-linux-x86-64.so.2 -> /lib/x86_64-linux-gnu/ld-2.24.so

Would be encoded like this outside the chroot:

    rootfs/lib64/ld-linux-x86-64.so.2 ->
    /fstemp/glass-9k4r6pu2/stage/rootfs/lib/x86_64-linux-gnu/ld-2.24.so

This is problematic, because when we package up the rootfs directory,
we expect the resulting package to be a valid root filesystem... but
it's not, because it has symlinks to paths that only existed during
the build process.

In order to fix this, we need an extra stage, after all of the
fakechroot invocations, which can strip out the garbage at the start
of each path.