chroot: add snapshot.debian.org cache
We rely on using snapshot.debian.org for reproducibility.
Unfortunately, it is also very slow, and this has a severe effect on
the speed of our build process. In order to improve build performance,
we need to cache downloaded debian packages in our upstream directory.

Add the nginx configuration required for this cache.