upstream: add debian package snapshot scripts
We rely on using snapshot.debian.org for reproducibility.
Unfortunately, it is also very slow, and this has a severe effect on
the speed of our build process. In order to improve build performance,
we need to cache downloaded debian packages in our upstream directory.

Add the scripts required to download the relevant subset of packages;
indirectly add the current cache to the upstream repository.