circleci: pull down matching upstream branches
The SHA512SUMS.UPSTREAM file needs to be in sync with the upstream/
directory. When processing pull requests, we might need newer changes
to upstream/, so let's specifically check out the branch to match the
branch that we're building for ... or, if there's no such branch, just
use the default.