spire: eliminate cdpack system
Remove the separate step of extracting the debian ISO during build.

The spire cdpack system was implemented when we didn't have bsdtar, and
correspondingly needed to perform an actual mount operation to read the
files out of the debian base ISO. We no longer need to do this, so the
cdpack process can be re-integrated back into the main ISO remastering
process.

We also relocate the debian ISO to /usr/share/spire/debian.iso, because
it doesn't make much sense to incorporate this kind of binary into the
deployed spire binary, when we could also just put it in the same
debian package.