cluster: bump container versions
Bump the versions of dnsmasq-nanny and flannel in the kubernetes
configuration files.

The previous commit bf84c1f4ba68f97c93e96dc002d0d7b79fcbb929 updated
the versions of the dnsmasq-nanny and flannel containers as a result of
updating their debian versions, but failed to also update the versions
used for cluster deploy itself.