upstream: associate repo by commit hash
We have an external upstream source code repository for the sake of not
cluttering our main repository with massive binaries. Originally, the
pull-upstream.sh script would simply grab the current branch,
defaulting to master. This was insufficient for CI, so we switched to
choosing the branch based on the current CI branch. This is also
insufficient, because it makes branches more meaningful than they
should be, and causes cross-contamination of CI validation between
branches, where a commit could fail on one build but succeed on another
because of the CI launches occurring from different branches.

Associate the upstream repository by commit hash, rather than branch.
This requires us to specifically update the hash file, but we have to
update SHA512SUMS.UPSTREAM anyway, so this is not a big deal. This
makes our CI builds more reliable and our upstream management process
more reasonable.

We don't use git submodules, because we want our pull script to verify
the hashes on the files received. There may be solutions to this
problem that involve submodules, but we haven't found one that works
well yet.