spire: add installer mirror configuration option
Hardcoding debian.csail.mit.edu is a bad idea, because it is not
guaranteed to always work, and because it may be across the country for
some users. This is especially relevant for local VM deployments, which
is the most likely reason to need a mirror located somewhere else.

Configure the mirror in setup.yaml, and interpolate it into the preseed
file.