registry: rename homeworld.mit.edu to homeworld.private
We serve our container meta discovery page from homeworld.mit.edu, and
this host is correspondingly hard-coded into the name of each built
container.

The domain 'homeworld.mit.edu' only works for the Hyades cluster, and
correspondingly should not be hardcoded. We also want to avoid needing
to recompile Homeworld for different clusters, so we need to use a host
that is portable to all clusters. Change to a hostname that isn't
publicly resolvable, so that each cluster can take it on itself to
redirect it to the appropriate container host.

Use 'homeworld.private' as the hostname, because the '.private' tld is
vaguely recommended per RFC 6762 appendix G. Avoiding a TLD entirely
appeared to cause problems with Kubernetes failing to recognize that
the containers had been correctly pulled, and therefore is not done.