install: populate hostname from keysystem
We implicitly rely on DNS being properly populated and available to
look up the correct hostnames for our nodes. We can't expect those
conditions to hold. Instead, set the hostname to a fixed value during
install, and populate it with the correct value once we've fetched the
node-specific configuration from the keyserver.