spire: populate kerberos realm directly
We previously let krb5-config populate the default kerberos realm
automatically based on the hostname derived from a reverse DNS lookup.
Now that we stop receiving the hostname from DNS, we need to specify
the kerberos realm in some other way. Preseed this value directly from
the setting in the cluster configuration file.