spire: allow renewing ssh certs without keygateway
If kerberos is disabled, we would have no way to request a signed ssh
cert from the keysystem, and wouldn't be able to establish ssh access
to the cluster. In this case, directly extract and employ the ssh
authority to generate a ssh cert ourselves.

This requires having the disaster recovery token available, which is
why this is not a security problem.