spire: add iso generation mode for serial install
When we want to automatically deploy a cluster, such as for CI, we need
the ability to perform the install without a graphical user interface.
Provide an option to generate an ISO that provides the isolinux boot:
prompt and debian install interface over serial.

At the boot prompt, a user (or program) could type a boot command like

    install netcfg/get_ipaddress=10.1.0.2 homeworld/asktoken=manual

and see their install complete, reporting over serial, without any
further intervention.