spire: allow admitting nodes without keygateway
If kerberos is disabled, we can't use the keygateway to request
bootstrap tokens. Use keyinitadmit in spire for fetching bootstrap
tokens in these cases.

keyinitadmit is an existing mechanism that allows getting a bootstrap
token for the supervisor node itself, through SSH to the supervisor
node. Expand keyinitadmit to support bootstrapping other nodes as well.