services: provide correct key to rkt
Previously, a fixed public key was provided to rkt for ACI signature
verification, preventing developers without the corresponding private
key from deploying the cluster. Here, we instead extract the key from
GPG, as is done in homeworld-apt-setup.

The implementation is rather kludgy since glass doesn't provide a native
way to write to a file with a dynamically determined name (rkt expects
the name of the file to be the fingerprint of the key). Any attempts at
generalization (e.g. modifying the build step) is deferred to when
similar features are required elsewhere.