glass: allow GOPATH customization
Previously, the only alternative to a fixed stagedir path for a
go-build block's GOPATH was specifying a different stagedir path.
Support specifying an alternate path as an absolute path, and support
specifying multiple path elements, rather than just one.

This could be used, for example, to include the building/src directory
directly in package search, as in:

  gopath: /h:go

(instead of the default gopath: go)