ci: use same process to mirror gpg keys in chroot
The original script in enter-chroot-ci.sh to copy GPG keys into the chroot
used pubring.gpg instead of pubring.kbx (which no longer works in the
latest version of GPG) and did not copy private keys in (which was
originally unnecessary since no apt repository was created in CI). This
change fixes both problems by using the same script as is used in the
non-CI enter-chroot.sh.

For compatibility with CircleCI, the code for the old version of GPG is
left in.