spire: eliminate list.yaml
The file list.yaml was used only to list possible cluster specs for a
debugging command. Remove it, and convert the command into a command
that generates a single spec.

The new user access grant system requires embedded special secrets that
require extra authority extraction code that we'd like to not have also
run for every other spec deployment. This means that the user grant
spec won't be able to be generated like the others, which means that
the spec debugging command would always fail to complete. Adjust the
debugging command so that this failure is limited to generating the
specific specification, by requiring its user to specify which
kubernetes cluster specification they want to have generated.