user-grant: add service for granting user access
We need a service to grant access to the kubernetes to users based on
certificates issued by a third party. Add an outline of such a service,
in the form of a single webpage that authenticates a user with TLS
client certs and gives them a kubeconfig file to use. Add scripts to
run this container on the cluster and provide the correct secrets and
configuration to run. Configure the expected hostname through the
cluster configuration file.