components: centralize source files to src/
We have a unified src/ tree for golang code, but this source tree
should also include the rest of our code. Move systemd unit files,
launch scripts, and other source code to the tree.

Rename some of the shell scripts to launch-*.sh so that they can be
easily recognized during the homeworld-services build.

This does change the packages responsible for installing some of the
files, but that should be okay.