spire: fix initrd creation in iso gen
Previously, the initrd GzipFile was directly set as the output of the cpio
subprocess. This causes subprocess to set the output of cpio to the file
descriptor of the GzipFile, which turns out to be the gz file itself.
This means that the data is written in uncompressed, because GzipFile does
the compression using an overriden write method. This change uses the
write method instead so that the data will be compressed.