glass, spire: Change format of apt branches.
The original motivation is to separate the upload and download processes
from each other, instead of using an ad-hoc scheme to try and guess the
upload path from the download path.  Now the download path is just a URL
which is simpler.  In addition, branch names are now independent of the
download URL, so you can name it something easy to remember.