kube-state-metrics: decontainerize
Launching kube-state-metrics through a container is unnecessary and
simply adds extra complexity. Strip out this extra layer of
indirection.

We now bind to the host ports directly, instead of binding to the
default ports and forwarding those from the host.

The way we generate the kubeconfig is still a bit sketchy, but it's no
worse than it was.

Add the new package as a dependency for homeworld-services, so that it
actually gets installed.