rkt: remove problematic rkt script
Because the versioning model was changed, it is now a known situation
to attempt to upgrade rkt while existing rkt containers are running.
The preinstall script prevents installation when this is the case;
bypass this interlock.

This might be a bad idea, but it's not clear what the failure mode is,
and there's a clear failure mode of the presence of this interlock.