bazel: import source code for go_repository
We're going to need an altered version of go_repository to handle some
of our upstream repositories correctly, so that we can perform patching
steps both before and after BUILD files are generated by Gazelle.

Make a copy of the original module to start with, unchanged from the
current bazel_gazelle upstream at version 0.15.0.