kubernetes: port to bazel
Port the Kubernetes build. Pull in a version of the kubernetes build
repository and a set of patches that will make this old version (1.9.2)
function correctly with the newer version of bazel.

Note that 0110-rerun-gazelle.patch, although ridiculously large, is
really just an autogenerated patch. See its header for regeneration
details. Carrying around a patch this large isn't great, but it's only
a temporary measure until we've upgraded kubernetes enough to no longer
need these patches -- or at least, to need fewer of them.

Note that we also still keep using the previous 0200-aci-launch.patch.