upload: implement basics of new system
Convert previous upload code in glass to new uploading framework.
Integrate with bazel. Structure as a build target that generates the
files ready to upload, and produce an executable that runs outside of
bazel to perform the actual upload.

To perform an upload:
    bazel run //upload

To build everything in preparation for an upload, except actually
uploading:
    bazel build //upload

One weakness of this system as it stands is that signing occurs during
the build process, and PGP passphrase pinentry conflicts in weird but
ultimately vaguely working ways with Bazel's output.

As before, a branches.yaml must be populated in the upload directory by
the local developer, and must not be committed into the actual project.

Instead of using the environment variable HOMEWORLD_APT_BRANCH to
specify the branch to build, we now place it in a special BRANCH_NAME
file in the upload folder, which means less work to specify branches in
the common case that the same branch is used repeatedly. This also
allows packages to be automatically rebuilt as necessary when the
branch name changes.