autostart: port to bazel
There was no original package to correspond to this, because all of the
services were scooped up by homeworld-services instead of their
individual packages, and this wasn't complicated enough to deserve its
own package.

Now that packages are lighter weight, add a package for this, declare a
dependency on it, and upload it.