dnsmasq: port to bazel
This C binary isn't trivial to rebuild ourselves, so we bypass that
complexity by directly shelling out to the original makefile, ensuring
that we do so in a temporary directory to avoid messing with the
original source tree.

Change pinning from key-based to commit-based, because it's easier to
do it this way with Bazel. Build an ACI, but don't upload it, because
it's only necessary as an intermediate step.