rkt: port to bazel
Download dependencies using bazel's tools, where we use pinned hashes
instead of signatures.

Because rkt's build process is incredibly complex, don't even try to
run it ourselves. Just use their makefiles directly. This is not a good
solution long-term, because it's slow and non-incremental, but it works
for now. Perform these builds in a separate directory to avoid
contaminating the source tree.

Construct a debian package similar to the one previously constructed
and upload it.