spire: port to bazel, rename from admin-tools
Port the spire command's package to build using Bazel. Add a bazel
macro to mark a zip file as executable with python.

Previously, we referred to the administrative tools both as "spire" and
as "homeworld-admin-tools" in different places. Work to unify this by
renaming the package to homeworld-spire as we port it.