spire: parameterize with debian version
Now that we update the versions on everything simultaneously, we can no
longer have versions hard-coded in various places in the code. Provide
the ability for service launchers in the cluster to parameterize based
on the current version. Also provide the ability for deployed
kubernetes specs to be parameterized by the current version.