circleci: fix continuous integration system
The CI configuration is still for the original glass build system.
Replace it with an appropriate configuration for bazel.

The bazel build process now requires signing the binaries. This means
that we can no longer avoid having a CI testing key. Hard-coding a key
doesn't solve this perfectly, but it solves it well enough, and because
the key is never used for anything important, this is fine.

We should probably consider replacing this with a key generation step,
though.

Also, we move the gpg import into the chroot itself, to avoid anything
weird happening with mismatched gpg versions between the host and the
chroot, and to avoid needing to specially copy around the gpg files.

We no longer need to have a step for verifying upstream dependencies.
Since the upstream repository is now phased out, a simple compile will
perform the checks we need. This is definitely more dependent on the
active availability of all of our upstream dependencies, but this
should hopefully be okay.