apt: add https support
The Jenkins web interface and local CI apt repository are served off the
same domain name. For security, HSTS is enabled for the domain, which
causes all http requests to redirect to https. This means that apt has to
work under https too, which just requires apt-transport-https to be
installed.