bazel: add OCI build macros
We need the ability to build containers as OCIs instead of ACIs. Based
on rules_docker, introduce a macro mostly compatible with homeworld_aci
that allows for the construction of an OCI.

OCIs contain no precoded version, but we still need a way to reference
particular ACI instantiations. We do this by hash instead. Expose a
.ocidigest file on each built OCI that contains the OCI image digest of
the built image. This can be used with the kubernetes digest reference
format:

    docker://homeworld.private/pullcheck@sha256:AAA....AAA

(the contents of the .ocidigest file would be the sha256:AAA....AAA
segment.)

As well, we include a .shortdigest file, which drops the initial
'sha256:' prefix. We'll need this when we're working around a kube+rkt
integration bug by using tag downloads instead of digest downloads, and
need a hash to use as a tag name -- which can't include colons.

For context, OCIs are the standardized container format from the Open
Container Initiative (https://www.opencontainers.org/), which appear to
be on track to replace Docker images and CoreOS ACIs.