registry: use docker registry instead of static
Background: ACI registries are easy, because they're a static HTML page
with an "ac-discovery" meta tag that points to a real location where
containers are stored.

OCI registries are harder to find. We use the docker registry, because
it seems to be the simplest and best-supported OCI-compliant registry.
Unfortunately, it is mutable, and needs containers to be pushed to it,
rather than simply serving out of a directory.

Pushing needs to be authorized, but the docker registry doesn't support
distinguishing between push access and pull access. Since we need to
let anyone pull without authentication, we add an nginx proxy layer,
which only allows GET requests, and only expose the raw API locally.

One of the docker registry's Go dependencies requires a certain library
on Go versions below 1.7, but that library is not included. This is
fine, but we have to patch the generated BUILD file to remove that
dependency from Bazel's build graph.

We also only support HTTPS image downloads, because there is no reason
that anyone should be trying to pull from this registry over HTTP, and
this helps avoid the possibility of someone accidentally doing that.

We also add a systemd service file to manage the docker registry, and
we enable it when we enable nginx.