registry: receive containers through deb package
The docker registry needs containers to be pushed to it, unlike the
previous static registry. Under normal conditions, we'd upload directly
from bazel with the oci_push rule, but the supervisor node for a
deployed cluster won't necessarily be running during the bazel upload
process, so we can't upload directly.

Transfer the containers to the supervisor with a set of debian
packages, each corresponding to one of the container images. Add an
update-registry command to spire, to push the containers from the
package to the registry, staying within the supervisor localhost.

Since there isn't any portable single-file format for OCIs, and only a
portable upload protocol, use a format based on the rules_docker
internals to store the containers. Upload out of this format using the
pusher binary from the containerregistry package, as would be used by
the oci_push rule.

Create an oci_to_tar rule to generate transferrable tar files of the
internal format in a form that can be incorporated into a debian
package, by smashing together the oci_push rule's code with the pkg_tar
rule's code. We can't use the pkg_tar rule directly, because it
requires knowing the list of files to include during the loading phase,
before oci_push has figured out what those are.

Unlike ACIs, OCIs have no inherent version, so we can just upload them
directly and recall them later by digest. Unfortunately, we also will
need to tag them by digest, due to a bug in the kubernetes + rkt
integration, so we also incorporate the digest so that we can upload to
the corresponding tag.