registry: switch downloads to docker protocol
Since we switched to running an OCI-compliant registry instead of a
static ACI registry, we can no longer download our containers using the
ACI autodiscovery protocol. Turn off our custom kubernetes option that
supports the ACI protocol, and add code to aci-pull-monitor to inform
rkt that it should be downloading over the OCI registry protocol, which
is also known as the docker image download protocol.