bazel: add hash_compute rule
In order to allow us to detect when the contents of a package changes,
and therefore when a version number needs to be automatically bumped,
we need to have an identifier that refers to the actual contents of the
built package.

Add a bazel rule hash_compute that produces a hash from a list of input
files and a list of strings.

In homeworld_deb and homeworld_aci, construct a hash based on all of
the files and attributes that could be varied by the invoker.

For now, we don't use these generated hashes, but we will.