bazel: add version calculation scaffold
We'll need code to calculate the version number to use for each
package, based on the global minimum version number and the hash of
that package.

Add a version_compute rule to wrap a python script that computes
version numbers. For the time being, simply use the global version
number and don't bother actually computing a local version.

Compute versions for acis and debs using this rule based on the hashes
computed previously.