bazel: compute version cache
In order to figure out when to bump version numbers, we're going to
need saved state that describes which versions have been uploaded, and
at which hashes.

Generate this state from the hash and version computed for each
package. Add an extra name target computed so that we can track the
name of each package as well. Store this state in the form of a JSON
object:

  {"homeworld-package": {"hash": "... hash ...", "version": "X.Y.Z"},
   ... }