debian-iso: bump version
The old ISO is no longer being hosted, so we need to switch to the new
version.

I reused the old upstream-check makefile to securely determine the new
hash. This implies that we still need some of the verification
functionality removed as part of the Bazel port, but it's going to have
to be in a new format, so we shouldn't keep that code around.