bazel: calculate incremental versions
Add a version cache file and use it to compute package versions.

The concept here is that, when a package changes hash, its version
should be incremented for the next upload. This allows us to only
re-upload binaries when they have actually changed -- without requiring
us to manually track which binaries we change between uploads.

The version cache is currently initialized empty and is never updated.