version: populate cache on upload
We don't want to bump the version number of a package every single time
it is modified; only when we have potentially installed the previous
version. So only bump versions at most once per upload.

Implement this by copying the generated version cache to the input
version cache at the start of each upload.