aci-pull-monitor: embed pullcheck version
The existing mechanism for aci-pull-monitor to fetch the correct
version of pullcheck involves spire embedding the version number into
the cluster configuration, uploading it to the keyserver, having each
node's keyclient pull down the version, and then reading it from
cluster.conf in a bash script invoked from the aci-pull-monitor systemd
service.

This is a large number of steps, and we now need to track the version
of the pullcheck container separately, because each component is
independently versioned, which is going to make spire even more
complicated.

Rip out the existing plumbing, and substitute the incremental version
for the pullcheck container directly into the systemd service file
instead.