registry: pull container images by digest
Using version numbers to track fixed versions of images isn't as
reliable as using hashes, and requires extra code to track the versions
to use for uploading. Pull images by digests instead, using the docker
pull-by-digest syntax.

However, due to a bug in the interaction between rkt, kubernetes, and
the digest syntax, we cannot actually do this properly in the
kubernetes specs. In that case, we instead use labels corresponding to
the non-prefix portion of the digest. (Which is a stop-gap solution.)

For reference, the two syntaxes:
  tag:     docker://homeworld.private/pullcheck:<TAG>
  digest:  docker://homeworld.private/pullcheck@sha256:<HASH>