runc: build and package
We're replacing rkt with something better supported. runc is the core
component of the replacement. It does not handle strong isolation like
rkt's kvm stage1, but we'll get to that later by integrating Kata
Containers.

Pull down the latest version of runc, build it, package it, and run it.

This commit does not yet enable runc's special build options, like
turning on seccomp.