cri-o: build and package
CRI-O is a tool to connect Kubernetes (via CRI) to runc (via OCI). This
tool will allow us to run containers with runc. Compile it and build a
package. Depend on runc.

Because CRI-O isn't pure go, we need to add bazel scripts to it to
build the C subprograms.

However, we need to make changes to the Go build process as well:
Gazelle doesn't handle #cgo pkg-config directives, which means that cgo
code that depends on them (such as to include glib-2.0) fail to compile
under Bazel. Address this by directly adding the relevant include and
link potions to the bazel scripts for those libraries.

Unfortunately, glib-2.0, which is widely required for some of CRI-O's
dependencies, requires a configuration file that is kept in the folder
/usr/lib/x86_64-linux-gnu/glib-2.0/include. This is outside of Bazel's
standard sandbox paths, which means that we can't just include a -I
directive, and we can't depend on the folder, because it's outside of
the workspace. So we add symlinks to that folder, and add cc_library
rules to pull in the glibconfig.h header from it.