cri-o: add null image signing policy
For the time being, we don't have image signing set up, so we need to
provide an empty image signing policy.json that will accept anything we
download from a docker registry.

This is fine, because we already ensure security by pinning container
image digests.

Add this through a dedicated package, because we need it to use skopeo,
which we'd like to not depend on the installation of CRI-O.