kubernetes: switch from rkt to cri-o
We previously used rkt to run containers under Kubernetes. This doesn't
make sense at this point, since rkt is not well-supported by the
community, and doesn't integrate well with Kubernetes anymore, as a
result of that loss of support. Replace our dependency on rkt with a
dependency on CRI-O + runc, which are significantly better-supported.

This does mean we can't use the kvm stage1 for isolation, but we'd
encountered issues with that anyway, and now we can introduce something
like Kata Containers to perform that function to replace runc.