cri-o: use pause from kubernetes
CRI-O originally built its own pause binary because it couldn't pull in
kubernetes/pause from the main registry. We need to build this binary
ourselves now, but CRI-O has a very old version of the pause code. Use
the newer version already available in the kubernetes source
distribution. Also, remember to use static linking, because this is
needed in a container.