spire: support foreach on kube nodes
It's often the case that we need to run a command on all nodes that run
kubernetes, but not the supervisor node. The current options of
"everything", "supervisor", "masters", and "workers" aren't sufficient,
because we end up having to run many commands through both masters and
workers. Add "kube" as a shortcut for every node but the supervisor.