oci-tools: build and package
We need the ability to test various elements of the container runtime
stack. Add the tools required to directly work with runc: skopeo to
download images and image-tools to unpack them into a runtime bundle.
Package these tools together and upload them, but don't include them in
any default dependencies, so they'll only be available if explicitly
installed for debugging purposes.

Some of the same build difficulties as CRI-O are addressed with code
copy-pasted from those patches.