pull-monitor: begin CRI-O support
We need to support testing container pulls and launches through CRI-O
in the pull monitor, instead of through rkt. Strip out the rkt support,
and replace it with a scaffolding for CRI-O support.

Skip over the exact details of how to talk to CRI-O for the moment, and
just use a set of TODO stubs. These will be replaced with real
implementations in the following two commits.

Replace the dependency on rkt with a dependency on CRI-O. Import some
additional error-handling libraries to make the code more
straightforward.