flannel: remove rkt stage1 annotation
Now that we don't use rkt, flannel should no longer specify which rkt
stage1 should be used. Remove the annotation.

Once we switch to Kata Containers, in the far future, we will need a
way to specify that flannel should run with runc instead of Kata.
According to the current documentation, our 'privileged' setting for
flannel should be sufficient for this, so we won't replace this with
another more specialized annotation or setting at this time.