apiserver: drop tls-ca-file parameter
This parameter is not actually used by modern versions of kubernetes,
so we no longer need to specify it. Note that the apiservers have no
need to connect outbound to anything but etcd and occasionally some
controllers -- and we don't use any of the latter.

This is required because kubernetes 1.11 will not support this option,
so we need to remove it before we upgrade.