keyclient: generate config in keysystem
It doesn't make sense for us to do this complex configuration
generation dance, where spire does template substitution to generate
the keyclient configuration, when the configuration is hardcoded in the
Homeworld source code anyway. Generate the configuration for the
keysystem in the keysystem itself, with the intention of eventually
removing the indirection through a separate configuration file
entirely.