kubernetes: load br_netfilter for proxy
The kube-proxy requires br_netfilter to work properly. If we don't have
it loaded, we get errors like the following when we try to make DNS
requests:

    ;; reply from unexpected source: 172.18.1.3#53, expected 172.28.0.2#53

Of course, br_netfilter is actually loaded by default by kubernetes up
through 1.10 -- but in 1.11, the compatibility code that does so is
removed as part of commit 37d30a081528f01ca554251f0f10539c9bcacad7.
Specifically, that commit removes the call to InitNetworkPlugin that
would invoke the NoopNetworkPlugin.Init method, which actually runs
"modprobe br-netfilter".

So, to be compatible with 1.11, we'll need to start loading this module
ourselves. This is actually done by kubeadm as well in kubernetes 1.13,
as added in commit c23e94a7ff17323c909619c3ba5e68e1c3f830ff -- but
since we are neither using kubeadm not kubernetes 1.13, this doesn't
immediately help us, and we've got to load it ourselves.