kubernetes: upgrade to 1.12.7
Bump the kubernetes version to a newer major release. This also means
that we're using a recent enough version of kubernetes that it
automatically works with the latest bazel, and we don't need to include
a bunch of backported patches for it. We're now on a stock build of
kubernetes!