bazel: don't commit stub version-cache in git
We don't want changes to the version cache to be tracked in git. We
already added version-cache to the .gitignore, but it's not enough,
because we need it to be populated before we invoke bazel, so we
included a stub version-cache in git.

Eliminate the stub version, and instead populate it when we enter the
build chroot if it doesn't already exist.