keyclient: always download public keys
We currently only download certain public keys onto a node if the
variant setting of the node signals that they'll be needed.

This is an unnecessary optimization: we should just always download
these. There's no reason we can't, because there's no authentication
for downloading public keys, and it simplifies the code.