keysystem: denote permission denied errors
Normally, we don't pass any information at all from the keyserver to
the keyclient about why a request failed. However, we'd like to have
special handling for what to do when a request is blocked because a
node doesn't have permission to run it, so start propagating
information about whether the error was a permission denial or not.