keyclient: request all keys regardless of variant
The last change made it possible for us to accept nominal-case
forbidden errors from the keyserver, because we'll simply wait a while
and not repeatedly retry anything that fails in such a manner.

This means that we no longer need to decide on our configuration based
on a variant sent by the keyserver in a previous run; we'll get what we
get, and that must correspond to what our variant is supposed to have.