keyclient: eliminate unnecessary variant tracking
Now that we don't change our actions based on the variant, the only
reason we need to keep track of variants is so that we know when to
update the hostname and CA certificates, because we usually do that at
the same time as switching away from the base variant.

This is a bad reason to include all of our variant tracking, and it's
not even clear that this is the right time to run these commands.

Eliminate the keyclient tracking of the node's variant, and replace the
script that would have restarted the keyclient with the new variant
with a script that will just directly set the hostname and update the
CA certificates.