keyclient: simplify actloop action interface
Start replacing the current boilerplate-heavy actloop structure with a
structure based on a single function entrypoint: the idea is that,
instead of an action being a collection of functions that get called in
different scenarios with simple interfaces, an action is a function
that gets called with a context object. This means that actions are
composable by just calling multiple action functions in a row.

For the moment, stop-gap this transition by adding a converter from an
old action structure to the new action entrypoint function.