keyclient: remove update-host script
Rather than having a special script that is dispatched with systemd
when the keyclient completes to update the hostname and CA
certificates, directly control the hostname and CA certificates from
special-case code in the keyclient. This is less messy, and since the
keyclient isn't planned to be used for any other projects, doesn't
limit us in any real way.